Drop `exit $(...)` stuff from file tests
These basically are a holdover from a long time ago. It appears they no
longer work properly. So this simply drops them. `test -f` appears to
actually do the right thing in all cases it has been used. So it should
be more than sufficient for performing the actual tests.